// src/apps/roles/emuns/role-messages.enum.ts
export enum RoleMessagesEnum {
  ROLE_NAME_INVALID = 'roles.name_invalid.error',
  ROLE_UUID_IS_INVALID = 'roles.role_uuid_invalid.error',
  NOT_FOUND = 'roles.not_found.error',
  ALREADY_EXIST = 'roles.already_exist.error',
  ROLE_INVALID = 'roles.invalid.error',
  key_INVALID = 'roles.key_invalid.error',
  ENABLED_INVALID = 'roles.enabled_invalid.error',
  PERMISSIONS_UUIDS_ARE_INVALID = 'roles.permissions_uuids_invalid.error',
  PERMISSION_KEY_INVALID = 'roles.permission_invalid.error',
  PERMISSIONS_NAME_INVALID = 'roles.permission_name_invalid.error',
  ACCOUNT_UUID_INVALID = 'roles.account_uuid_invalid.error',
}